+++ indico_toolkit/etloutput/__init__.py
@@ -1,13 +1,13 @@
 from typing import TYPE_CHECKING, TypeAlias, TypeVar
 
-from ..results import NULL_BOX, NULL_SPAN, Box, Span
-from ..results.utils import get, has, json_loaded, str_decoded
-from .cell import Cell, CellType
-from .errors import EtlOutputError, TableCellNotFoundError, TokenNotFoundError
+from .box import NULL_BOX, Box
+from .cell import NULL_CELL, Cell, CellType
 from .etloutput import EtlOutput
-from .range import Range
-from .table import Table
-from .token import Token
+from .range import NULL_RANGE, Range
+from .span import NULL_SPAN, Span
+from .table import NULL_TABLE, Table
+from .token import NULL_TOKEN, Token
+from .utils import get, has, json_loaded, str_decoded
 
 if TYPE_CHECKING:
     from collections.abc import Awaitable, Callable
@@ -17,17 +17,18 @@ __all__ = (
     "Cell",
     "CellType",
     "EtlOutput",
-    "EtlOutputError",
     "load",
     "load_async",
     "NULL_BOX",
+    "NULL_CELL",
+    "NULL_RANGE",
     "NULL_SPAN",
+    "NULL_TABLE",
+    "NULL_TOKEN",
     "Range",
     "Span",
     "Table",
-    "TableCellNotFoundError",
     "Token",
-    "TokenNotFoundError",
 )
 
 Loadable: TypeAlias = "dict[str, object] | list[object] | str | bytes"