+++ indico_toolkit/etloutput/token.py
@@ -1,7 +1,9 @@
 from dataclasses import dataclass
+from typing import Final
 
-from ..results import Box, Span
-from ..results.utils import get
+from .box import NULL_BOX, Box
+from .span import NULL_SPAN, Span
+from .utils import get
 
 
 @dataclass(frozen=True)
@@ -9,6 +11,9 @@ class Token:
     text: str
     box: Box
     span: Span
+
+    def __bool__(self) -> bool:
+        return self != NULL_TOKEN
 
     @staticmethod
     def from_dict(token: object) -> "Token":
@@ -23,3 +28,14 @@ class Token:
             box=Box.from_dict(get(token, dict, "position")),
             span=Span.from_dict(get(token, dict, "doc_offset")),
         )
+
+
+# It's more ergonomic to represent the lack of tokens with a special null token object
+# rather than using `None` or raising an error. This lets you e.g. sort by the `token`
+# attribute without having to constantly check for `None`, while still allowing you do
+# a "None check" with `bool(extraction.token)` or `extraction.token == NULL_TOKEN`.
+NULL_TOKEN: Final = Token(
+    text="",
+    box=NULL_BOX,
+    span=NULL_SPAN,
+)