+++ indico_toolkit/results/__init__.py
@@ -1,13 +1,10 @@
 from typing import TYPE_CHECKING, TypeAlias, TypeVar, overload
 
+from ..etloutput import NULL_BOX, NULL_SPAN, Box, Span
 from .document import Document
-from .errors import ResultError
 from .predictionlist import PredictionList
 from .predictions import (
-    NULL_BOX,
     NULL_CITATION,
-    NULL_SPAN,
-    Box,
     Classification,
     DocumentExtraction,
     Extraction,
@@ -15,7 +12,6 @@ from .predictions import (
     FormExtractionType,
     Group,
     Prediction,
-    Span,
     Summarization,
     Unbundling,
 )
@@ -45,7 +41,6 @@ __all__ = (
     "Prediction",
     "PredictionList",
     "Result",
-    "ResultError",
     "Review",
     "ReviewType",
     "Span",