+++ indico_toolkit/results/predictions/formextraction.py
@@ -1,16 +1,14 @@
 from dataclasses import dataclass
 from enum import Enum
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
-from ..review import Review
+from ...etloutput import Box
 from ..utils import get, has, omit
-from .box import Box
 from .extraction import Extraction
 
 if TYPE_CHECKING:
-    from typing import Any
-
     from ..document import Document
+    from ..review import Review
     from ..task import Task